chore: move functions to provenance extractor; perform provenance comparison on PURLs without versions for repo part
